# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Luxe Clothes Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a guest order (no login required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{orderNumber}:
    get:
      operationId: getOrderByNumber
      tags: [orders]
      summary: Get an order by its order number (confirmation page)
      parameters:
        - name: orderNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Admin Auth ────────────────────────────────────────────────────
...
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin-orders]
      summary: List all orders (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [processing, delivered, cancelled]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated admin orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderList"
  /admin/orders/{id}:
    get:
      operationId: adminGetOrder
      tags: [admin-orders]
      summary: Get a single order (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders/{id}/status:
    patch:
      operationId: adminUpdateOrderStatus
      tags: [admin-orders]
      summary: Update order status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Admin Coupons ─────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Settings ─────────────────────────────────────────────────────
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
      required: [id, productId, productName, quantity, unitPrice, totalPrice]
    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
        size:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
      required: [productId, quantity]
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerName:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        notes:
          type: ["string", "null"]
        subtotal:
          type: number
        deliveryFee:
          type: number
        discount:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [processing, delivered, cancelled]
        couponCode:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
      required: [id, orderNumber, customerName, phone, address, city, subtotal, deliveryFee, discount, total, status, items, createdAt]
    OrderInput:
      type: object
      properties:
        customerName:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        notes:
          type: ["string", "null"]
        couponCode:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
          minItems: 1
      required: [customerName, phone, address, city, items]
    OrderStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [processing, delivered, cancelled]
      required: [status]
    AdminOrderList:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [orders, total, page, limit, totalPages]

    # ── Admin / Auth ──────────────────────────────────────────────────
...